rbd: modify `ensureImageCleanup()` to be best effort
The function `ensureImageCleanup()`, called when images to
be deleted are not found, ensures that these images if in
trash have a task on them to be removed from trash.

It is an expensive operation to list images in trash depending
on the number of images in trash and this scenario is an
extreme edge case, therefore this commit modifies code
to only log the errors instead of returning it which
causes a retry of the whole request.

refer: #3593

Signed-off-by: Rakshith R <rar@redhat.com>