ci: make minikube.sh work on macOS M1 with the qemu2 driver
Some modifications have been made to handle the compatibility of bash scripts.
For example, `sed -i` without a value only works for GNU sed, but not for
macOS's sed.
The RESOLV_CONF variable has been made configurable because it needs to be
overridden as /etc/resolv.conf when using the qemu2 driver.
In addition, minikube also supports adding extra disks for the qemu2 driver.

Signed-off-by: zjx20 <zhoujianxiong2@gmail.com>