rbd: add additional space for encrypted volumes
issue: when a block-mode pvc is created with encryption enabled
there is some space reserved for the encryption metadata.
Which doesn't allows users to write extact amount of data that
they have requested for.

solution: create pvc with extra space needed for the encryption
metadata.

The extra space is added during the CreateVolume and ExpandVolume
operations. And while returning the response remove the extra
space so the client/user gets the requested size reported.

Signed-off-by: Praveen M <m.praveen@ibm.com>