ci: add e2e-build-test for compiling the e2e testsuite
When Go modules get updated, golangci-lint sometimes fails with weird
errors. One of the common causes seems to be that there is a dependency
breakage between modules that are only used within the e2e test suite. A
normal build of the cephcsi executable succeeds, but building ./e2e
would fail.

By adding a job to build the e2e.test executable, a clear error message
will be reported when there are package dependency conflicts.

Signed-off-by: Niels de Vos <ndevos@ibm.com>