build: reinstall the ceph-release package
With the new Ceph container-build process, the .repo files for DNF are
removed. This means that the build containers can not install required
dependencies anymore, like librados-devel and librbd-devel.

By reinstalling the ceph-release package from the Ceph distribution
server, the .repo files are restored and the required Ceph packages can
be installed again.

Signed-off-by: Niels de Vos <ndevos@ibm.com>