rbd: return gRPC code `Aborted` when the RBD-image is in-use on delete
According to the error scheme documented in the CSI specification, the
Aborted error code should be initiate retries, whereas the Internal
error code does not require this behaviour.

When an RBD-image is still in-use, it can not be removed. The
DeleteVolume procedure should be retried and will succeed once the
RBD-image is not in-use anymore.

Fixes: #5166
Signed-off-by: Niels de Vos <ndevos@ibm.com>