rebase: re-reun `go mod tidy`
The e2e/go.mod has been updated wih a newer Kubernetes version. Now `go
mod verify` complains about the need to run `go mod tidy` in the root of
the project.

It seems the previous Kubernetes update in the root of the project did
not update and vendor everything for some reason?

Fixes: 337d9a3 "rebase: bump k8s.io/kubernetes in the k8s-dependencies group"
Signed-off-by: Niels de Vos <ndevos@ibm.com>