cephfs: store clientaddress metadata to snapshot-backed volume
Problem: For a snapshot-backed PVC, there is no actual subvolume in the
Ceph cluster. Attempting to store metadata would fail as the decoded
subvolume name is not found in the cluster.

Solution: For a snapshot-backed PVC, store the metadata in its snapshot
using the volumeHandle to uniquely identify it. The clientaddress
metadata format is now `.cephfs.csi.ceph.com/clientaddress/<volId>/<nodeId>`.

Signed-off-by: Praveen M <m.praveen@ibm.com>