util: rename VolumeLocks to IDLocker as it is not limited to volumes
The VolumeLock structs and NewVoluemLocks() function provide a locking
API that isn't restricted to volumes only. The implementation is very
generic, and can be used by other parts of the code that need to lock
operations based on others resources.

By renaming VolumeLocks to IDLocker (it is already in the idlocker.go
file), the genericness of the locking API is expressed better.

Users of the API are recommended to call their IDLocker object with an
understandable name. VolumeLocks for locking resources related to
volumes, HostLocks for locking resources related to hosts, and so on.

Signed-off-by: Niels de Vos <ndevos@ibm.com>