e2e: add NVMe-oF gateway deployment
Rook is not yet able to deploy the NVMe-oF gateway, so it needs to be
done manually.

When the gateway starts, it uses a pool for fetching/storing some
metadata. If the pool does not exist, the gateway fails to start.
Before the gateway is deployed the `nvmeofpool` is created (and deleted
after testing).

Signed-off-by: Niels de Vos <ndevos@ibm.com>