e2e: auto-detect running on OpenShift
The e2e suite can run on OpenShift, but there are a few checks that
handle certain things with a small change. OpenShift has a more struct
security policy than standard Kubernetes/Minikube, these need to be
accounted for.

There used to be a -is-openshift commandline flag. This flag is now
removed and OpenShift is auto-detected.

Signed-off-by: Niels de Vos <ndevos@ibm.com>