ci: drop unneeded `test_type-` prefix from trigger phrase
Running CI jobs for a single storage type is now a little easier. Others
(and I) always need to check how the format for specifying the test_type
is. With this change, it becomes as simple as adding `/rbd` to the
trigger phrase.

Example: /test ci/centos/mini-e2e/k8s-1.34/rbd
Signed-off-by: Niels de Vos <ndevos@ibm.com>