ci: correct regular expression for triggering jobs with test-type
The `{test_type}` in the trigger phrase expanded to an array notation
like `['cephfs', 'nfs', 'nvmeof', 'rbd']`. This isn't a valid regular
expression, so passing the test-type on the trigger phrase could not
work.

By explicitly constructing the regular expression with all the available
test-types, this should now work (again?).

Signed-off-by: Niels de Vos <ndevos@ibm.com>