rbd:  add configuration option to set the cipher suite
Problem: When using RBD encryption it is only possible to use AES-XTS to encrypt an RBD image. Because in the current dmcrypt integration we only use the default.

Solution: I added the encryptionCipher and encryptionKeySize Option to the RBD storage class definition. Those options if selected will be passed down to dmcrypt encrypt the disk according to the specific cipher suite.

resolves ceph#5686

Signed-off-by: David Mohren <david.mohren@clyso.com>