rbd: introduce an allowlist for cipher and integrity options
Problem: Cipher and integrity option will be passed to the command line to call cryptsetup.
If user chooses a malicious string it can become a security problem.
Cryptsetup is called on the worker node and could be used to for example delete or copy files on the worker node.

Solution: Implement an allowlist for all cipher and integrity options. Hence if user passes
malicious string that is not in the allowlist it will not be passed to the command line.

Signed-off-by: David Mohren <david.mohren@clyso.com>