rbd: add e2e helpers for PV encryption state validation
New e2e tests need to compare the desired encryption options (the input to 'cryptsetup luksFormat') with the actual, on-disk state (the output of 'cryptsetup status').

These two are not a 1:1 match. This commit introduces a new 'LuksStatus' struct that translates the 'status' output into a format that can be compared against the 'EncryptionOptions'.

For example a problem was the is the keysize translation:

- 'cryptsetup luksFormat --key-size' ('encryptionKeySize' input) refers *only* to the data encryption key size.

- 'cryptsetup status' (output) reports a 'keysize' that is the *sum* of the data key + the integrity/HMAC key.

The new .Equal() comparator correctly accounts for this by subtracting the 'integrityKeysize' from the total 'keysize' before comparison.

Signed-off-by: David Mohren <david.mohren@clyso.com>